+++ wolf3dredux_0.01/gnulib_replacement_funcs/strnlen1.c
@@ -28,9 +28,9 @@
 size_t
 strnlen1 (const char *string, size_t maxlen)
 {
-  const char *end = (const char *) memchr (string, '\0', maxlen);
+  const char *end = (const char *)memchr(string, '\0', maxlen);
   if (end != NULL) {
-    return end - string + 1;
+    return (size_t)(end - string + 1);
   } else {
     return maxlen;
   }
